# apps/api/openapi-snapshot.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MediaCompressor API",
    "version": "0.1.0",
    "description": "REST API for MediaCompressor — image/video compression service"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mc_session"
      }
    },
...
  },
  "paths": {
...
    "/invites/{token}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 8
            },
            "in": "path",
            "name": "token",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 8
            },
            "in": "path",
            "name": "token",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/v1/admin/invites": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "expiresInHours": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 168,
                    "default": 24
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "get": {
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/admin/invites/{id}": {
      "delete": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/v1/internal/uploads/hooks/pre-create": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "Type": {
                    "type": "string"
                  },
                  "Event": {
                    "type": "object",
                    "properties": {
                      "Upload": {
                        "type": "object",
                        "properties": {
                          "ID": {
                            "type": "string"
                          },
                          "Size": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "MetaData": {
                            "type": "object",
                            "additionalProperties": {
                              "type": "string"
                            },
                            "default": {}
                          }
                        },
                        "required": ["Size"],
                        "additionalProperties": false
                      }
                    },
                    "required": ["Upload"],
                    "additionalProperties": false
                  }
                },
                "required": ["Event"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/admin/invites": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": ["created", "revoked", "csrf-stale"]
            },
            "in": "query",
            "name": "updateflash",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ]
}